Enhance: add qmake project files
The project that I wish to use GdbCrashHandler in provides both qmake and
cmake build systems so we need qmake project files as well. This PR
attempts to provide them.

It was necessary to modify the header inclusion for the QuaZip library for
the `./lib/GdbCrashHandlerDialog.cpp` file, it seems that Debian Linux, at
least, puts it in a `quazip` sub-directory.

I also note that this library is at version 0.3.0 and that is what I
specified in the QMake project file - so it build the chain of symbolic-
links to the un-versioned `libGdbCrashHandler.so` using those numbers.
However the Cmake project file produces version numbers based on a 1.0.0
semantic version number - I guess one of these is not right but it isn't
precisely clear yet why a sub-unity version number is being used to produce
a unity product version.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>